Add files via upload
I will be explaining the task that I performed in python programming language. In this problem, we are provided with the dataset and we have to create a linear regression model in Python or R to predict the CTC/Salary of new hires.

My task is to help John from STA IT services to create a linear regression model in Python or R to predict the CTC/Salary of new hires from the data provided.